requirements.txt
streamlit>=1.33.0
pillow>=10.2.0
numpy>=1.24.3
pandas>=2.0.1
scikit-learn>=1.2.2
torch>=2.1.0
joblib>=1.3.2
lightgbm>=4.3.0
matplotlib>=3.7.1
scipy>=1.10.1
xgboost>=2.0.3
openpyxl==3.1.2
torchcam
keras
tensorflow
scikit-image